do not break backwards compatibility
changing the default behavior of bump2version to always use the
search/replace functionality to update the version in the cfg-files,
brings several backwards incompatible changes:

    1. when current_version did not exist in .bumpversion.cfg, it will not be updated
     - when retrieving the version from the git tag
     - when retrieving the version via --current_version flag
    2. when .bumpversion.cfg contains a "new_version" field, previously this
      field was removed, but with this change it is not possible anymore

To get the desired functionality, we keep the current behavior.
But now, if the config-file is explicitely specified to be handled by bump2version search/replace, it is not rewritten.

The reason, why someone wants this is, that the old behavior of
rewriting the .cfg file discards all comments and custom formatting.